new function: mortality, conNeighbors, getNames, matToTabular
updated mortality function,
getNames function is an alternative to the scenario option in the params. But, it is less susceptible to errors, because the params are taken after the run.
included the getConNeigh function to get a matrix with the number of conspecific N. Other helper functions were implemented for this.
 Function to convert matrix information into tabular inforamtion is implemented.

all functions give all informations for every generation. No matter if the generations are directly consecutive or further apart.

the docu must be refined !
maybe at the end when other functions are ready too.