Create getParamNames.Rd
I had a bug(?): i changed a function in the R filder and rebuild the package. However the changes were not present in the rebuilded version, which kept the old version of the function. I spent a day trying to fix it. I suppose, as the function is present in sifferent branches, somehow it takes the old - not updated function from another branch, ignoring the new changes.

I bypassed the issue by building a new function, even though this is not the ideal solution